jwt integrated

+++ src/hooks/useToken.ts
@@ -1,4 +1,5 @@
 import { useEffect, useState } from "react";
+import axios from 'axios';
 
 export type TEmail = {
   email: string;
@@ -8,17 +9,25 @@ const useToken = (email: TEmail) => {
   const [token, setToken] = useState("");
 
   useEffect(() => {
-    if (email) {
-      fetch(`http://localhost:5000/jwt?email=${email}`)
-        .then((res) => res.json())
-        .then((data) => {
+    const fetchToken = async () => {
+      try {
+        if (email && email.email) { // Check if email and email.email exist
+          const response = await axios.get(`http://localhost:5000/jwt?email=${email.email}`);
+          const data = response.data;
           if (data.accessToken) {
             localStorage.setItem("accessToken", data.accessToken);
             setToken(data.accessToken);
           }
-        });
-    }
+        }
+      } catch (error) {
+        console.error('Error fetching token:', error);
+      }
+    };
+
+    fetchToken(); // Call the function immediately
+
   }, [email]);
+
   return [token];
 };
